+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 10.0.0-beta03
+- Allow setEnvironment({required bool useSandbox}) to enable sandbox or production environment
+- Allow setCallbackUrl({required Uri callbackUrl}) to set a callback url for all submitted jobs.
+- Bug Fixes and Improvements from iOS v10.0.0-beta10 and Android v10.0.0-beta09
+
 # 10.0.0-beta02
 - Support for Document Verification, exposed as a `SmileIDDocumentVerification` Widget
 - Support for SmartSelfie Authentication, exposed as a `SmileIDSmartSelfieAuthentication` Widget